==> 2. Data_Prep - 1 (SharePrice & Newz Data Merged).ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
from pandas import Series
#from mpl_toolkits.basemap import Basemap
import warnings
warnings.filterwarnings('ignore')
%matplotlib inline

#Reading in data and looking at the number of rows and columns using shape 
aapl_newz = pd.read_csv('AAPL_Newz_Data.csv')
aapl_newz.shape

aapl_price = pd.read_excel('AAPL_Openclose.xlsx')
aapl_price.shape

aapl_price.head()

#removing data less than 1st Jan 2015
aapl_price = aapl_price.drop(aapl_price.index[aapl_price['Date'] <= "2015-01-01"])
aapl_price.head()

aapl_price.shape

def diff_nxt_row(df, col, diff_col):
    ls = df[col].tolist()
    n = len(ls)
    diff_ls = []
    i = -1
    for row in range(0, n):
        i += 1
        if i < (n-1):
            diff_ls.append(ls[i+1] - ls[i])
        else:
            diff_ls.append(None)
    df[diff_col] = diff_ls     

diff_nxt_row(aapl_price, "Open", "fut_change_stock_price")
aapl_price = aapl_price.dropna()

aapl_price["Pos/Neg_fut_Change"] = aapl_price["fut_change_stock_price"].apply(lambda x: 1 if x>0 else 0)

#To check if data has been loaded properly
aapl_newz.head()

#removing null values from Date columns & changing "hours ago" to date
aapl_newz = aapl_newz.drop(aapl_newz.index[aapl_newz['Date'].isnull()])
aapl_newz = aapl_newz.replace(["hours 13 ago","hours 21 ago"],"31-Dec-21")

aapl_newz['Date'] = pd.to_datetime(aapl_newz['Date'])
aapl_newz.dtypes

aapl_newz.head()

aapl_combined = aapl_price.merge(aapl_newz, how = 'left', left_on = 'Date', right_on = 'Date')
aapl_combined.shape

aapl_combined.isnull().sum()

aapl_combined = aapl_combined.dropna()

print(aapl_combined.isnull().sum())
print(aapl_combined.shape)

aapl_combined.to_csv("AAPL_Combined.csv", index = False)

#conda install gensim
import gensim
from gensim import corpora

#Preparing Document-Term Matrix

# Creating the term dictionary of our corpus, where every unique term is assigned an index. 
dictionary = corpora.Dictionary(doc_clean)

# Converting list of documents (corpus) into Document Term Matrix using dictionary prepared above.
doc_term_matrix = [dictionary.doc2bow(doc) for doc in doc_clean]

"""Run LDA model"""

#Running LDA Model
# Creating the object for LDA model using gensim library
Lda = gensim.models.ldamodel.LdaModel

# Running and Training LDA model on the document term matrix.
ldamodel = Lda(doc_term_matrix, num_topics=10, id2word = dictionary, passes=50)

"""Print 10 topics"""

print("Topic 1")
print(ldamodel.print_topics(num_topics=10, num_words=50)[0])
print("Topic 2")
print(ldamodel.print_topics(num_topics=10, num_words=50)[1])
print("Topic 3")
print(ldamodel.print_topics(num_topics=10, num_words=50)[2])
print("Topic 4")
print(ldamodel.print_topics(num_topics=10, num_words=50)[3])
print("Topic 5")
print(ldamodel.print_topics(num_topics=10, num_words=50)[4])
print("Topic 6")
print(ldamodel.print_topics(num_topics=10, num_words=50)[5])
print("Topic 7")
print(ldamodel.print_topics(num_topics=10, num_words=50)[6])
print("Topic 8")
print(ldamodel.print_topics(num_topics=10, num_words=50)[7])
print("Topic 9")
print(ldamodel.print_topics(num_topics=10, num_words=50)[8])
print("Topic 10")
print(ldamodel.print_topics(num_topics=10, num_words=50)[9])
