==> 2. Data_Prep - 3 (Train-Test-Split).ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
from pandas import Series
#from mpl_toolkits.basemap import Basemap
import warnings
warnings.filterwarnings('ignore')
%matplotlib inline

#Reading in data and looking at the number of rows and columns using shape 
aapl_train = pd.read_csv('AAPL_Train.csv')
aapl_test = pd.read_csv('AAPL_Test.csv')

print(aapl_train.shape)
print(aapl_test.shape)

aapl_train.head(3)

aapl_test.head(3)

x_train, y_train, x_test, y_test = aapl_train['Top_Headlines_words'], aapl_train['Pos/Neg_nextday_Change'], aapl_test['Top_Headlines_words'], aapl_test['Pos/Neg_nextday_Change'] 

x_train.to_csv('x_train.csv', index = False)
y_train.to_csv('y_train.csv', index = False)
x_test.to_csv('x_test.csv', index = False)
y_test.to_csv('y_test.csv', index = False)
