==> 2. Data_Prep - 4  (TFIDF).ipynb <==
import pandas as pd
import numpy as np

#Reading in data and looking at the number of rows and columns using shape 
x_train = pd.read_csv('x_train.csv')
y_train = pd.read_csv('y_train.csv')
x_test = pd.read_csv('x_test.csv')
y_test = pd.read_csv('y_test.csv')
print(x_train.shape)
print(y_train.shape)
print(x_test.shape)
print(y_test.shape)

y_train.head(2)

x_train.head(2)

from sklearn.feature_extraction.text import TfidfVectorizer

# Converting the x_train and x_test dataframe to list data structure. As TfidfVectorizer doesn't work with dataframe. 
Headlines_Train = x_train['Top_Headlines_words'].tolist()
Headlines_Test = x_test['Top_Headlines_words'].tolist()
Headlines_Train[0:3]

# Setting up the parameters for TfidfVectorizer
tfidf_vec = TfidfVectorizer(max_features = None, max_df = 0.8, min_df = 5, ngram_range = (1,3))

# Fitting the TfidfVectorizer on Train Data
tfidf_vec.fit(Headlines_Train)

# Transforming Train Fit TfidfVectorizer on on both and x_train and x_test Data.
train_set = tfidf_vec.transform(Headlines_Train)
test_set = tfidf_vec.transform(Headlines_Test)

# Models:

## 1. Random Forest Classifier (RFC)

from sklearn.ensemble import RandomForestClassifier

# Matrix of 1472 by 3748 is made using TF-IDF
train_set

# Fitting the RandomForest Regression model on Train Data.
RFC = RandomForestClassifier(n_estimators = 100, min_samples_split = 2, min_samples_leaf = 5, max_features = 'sqrt', max_depth = 10, random_state = 42)
RFC.fit(train_set, y_train['Pos/Neg_nextday_Change'])

# R Sqaure of model on both Train and Test sets.
print("Random Forest Classifier")
print("R-Sqaure")
print(" ")
print("R-Sqaure on Train Set = "+str(np.round(RFC.score(train_set, y_train), 2)))
print(" ")
print("R-Sqaure on Test Set = "+str(np.round(RFC.score(test_set, y_test), 2)))
print(" ")

newz = 'wikileaks say release file cia cyber spying tool'

pred = RFC.predict(tfidf_vec.transform([newz]))
pred

from sklearn.tree import DecisionTreeClassifier
clf_model = DecisionTreeClassifier(criterion="gini", random_state=45,max_depth=10, min_samples_leaf=2)   
clf_model.fit(X_train_scaled,Y_train)

Newz_Senti.__dict__

#### RFC is overfitting

# Checking Accuracy
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score

predictions_rfc = RFC.predict(test_set)

matrix = confusion_matrix(y_test, predictions_rfc)
print(matrix)

acc_score = accuracy_score(y_test, predictions_rfc)
print(np.round(acc_score, 2))

report = classification_report(y_test, predictions_rfc)
print(report)

#### RFC Accuracy is 59%

## 2. Naive Bayes

from sklearn.naive_bayes import MultinomialNB

MNB = MultinomialNB()
MNB.fit(train_set, y_train['Pos/Neg_nextday_Change'])

# R Sqaure of model on both Train and Test sets.
print("Multinomial Naive Bayes")
print("R-Sqaure")
print(" ")
print("R-Sqaure on Train Set = "+str(np.round(RFC.score(train_set, y_train), 2)))
print(" ")
print("R-Sqaure on Test Set = "+str(np.round(RFC.score(test_set, y_test), 2)))
print(" ")

newz = 'ios'

pred = MNB.predict(tfidf_vec.transform([newz]))
pred

pred = RFC.predict(tfidf_vec.transform([newz]))
pred

predictions_mnb = MNB.predict(test_set)

matrix = confusion_matrix(y_test, predictions_mnb)
print(matrix)

acc_score = accuracy_score(y_test, predictions_mnb)
print(np.round(acc_score, 2))

report = classification_report(y_test, predictions_mnb)
print(report)

from sklearn.ensemble import GradientBoostingClassifier

GBoost = GradientBoostingClassifier()
GBoost.fit(train_set, y_train['Pos/Neg_nextday_Change'])

# R Sqaure of model on both Train and Test sets.
print("Multinomial Naive Bayes")
print("R-Sqaure")
print(" ")
print("R-Sqaure on Train Set = "+str(np.round(GBoost.score(train_set, y_train), 2)))
print(" ")
print("R-Sqaure on Test Set = "+str(np.round(GBoost.score(test_set, y_test), 2)))
print(" ")
